wow, you have to name them test

+++ tinygrad/gradcheck.py
@@ -8,7 +8,7 @@ def jacobian(func, input):
 
   ji = input.data.reshape(-1).shape[-1]
   jo = output.data.reshape(-1).shape[-1]
-  J = np.zeros((jo,ji))
+  J = np.zeros((jo,ji), dtype=np.float32)
 
   for o in range(jo):
     # tinygrad doesn't support slicing, tiny-hack to select
@@ -25,7 +25,7 @@ def numerical_jacobian(func, input, eps = 1e-6):
 
   ji = input.data.reshape(-1).shape[-1]
   jo = output.data.reshape(-1).shape[-1]
-  NJ = np.zeros((jo, ji))
+  NJ = np.zeros((jo, ji), dtype=np.float32)
 
   for o in range(jo):
     for i in range(ji):
